fix: consolidate agypty and provider packages for CI file-count limits
Merge agypty platform and shared files to stay within the 15-file package cap, move executeAgy into inference_progress.go, and relocate the Agy execute test to inference_provider_streaming_test.go so backend-size and pkg-maint gates pass.
