+++ pkg/workers/agypty/allocator_common.go
@@ -6,6 +6,11 @@ import (
 	"strings"
 )
 
+// ptyAllocation is the platform-specific PTY resource bundle returned by Allocate.
+type ptyAllocation interface {
+	Close() error
+}
+
 // PTYKind identifies which platform PTY mechanism was allocated.
 type PTYKind int
