--- pkg/workers/agypty/factory_unix.go
@@ -1,7 +1,0 @@
-//go:build linux || darwin
-
-package agypty
-
-func newPlatformPTYAllocator() PTYAllocator {
-	return NewPOSIXPTYAllocator()
-}
