--- pkg/workers/agypty/factory_unsupported.go
@@ -1,7 +1,0 @@
-//go:build !windows && !linux && !darwin
-
-package agypty
-
-func newPlatformPTYAllocator() PTYAllocator {
-	return NewUnsupportedPTYAllocator()
-}
