--- pkg/workers/agypty/factory_windows.go
@@ -1,7 +1,0 @@
-//go:build windows
-
-package agypty
-
-func newPlatformPTYAllocator() PTYAllocator {
-	return NewWindowsConPTYAllocator()
-}
