+++ pkg/workers/agypty/platform_unsupported.go
@@ -15,3 +15,21 @@ func NewUnsupportedPTYAllocator() PTYAllocator {
 func (unsupportedPTYAllocator) Allocate(context.Context, ProcessLaunch, SessionConfig) (PTYSession, error) {
 	return nil, ErrUnsupportedPlatform
 }
+
+func newPlatformPTYAllocator() PTYAllocator {
+	return NewUnsupportedPTYAllocator()
+}
+
+func runPlatformSession(context.Context, *platformSession) (SessionResult, error) {
+	return SessionResult{}, ErrUnsupportedPlatform
+}
+
+func sessionProcessRunning(int) bool {
+	return false
+}
+
+func terminateSessionTestProcess(int) {}
+
+func (p *sessionProcess) Wait() error {
+	return ErrUnsupportedPlatform
+}
