--- pkg/workers/agypty/pty_allocation.go
@@ -1,6 +1,0 @@
-package agypty
-
-// ptyAllocation is the platform-specific PTY resource bundle returned by Allocate.
-type ptyAllocation interface {
-	Close() error
-}
