+++ pkg/workers/agypty/session_run.go
@@ -5,9 +5,93 @@ import (
 	"errors"
 	"io"
 	"os"
+	"os/exec"
 	"sync"
 	"time"
+
+	"github.com/portpowered/infinite-you/pkg/workers/process"
 )
+
+// platformSession holds an allocated platform PTY for one supervised Agy child.
+// Story 002 implements Run with capture, timeout, and cleanup.
+type platformSession struct {
+	launch ProcessLaunch
+	cfg    SessionConfig
+	kind   PTYKind
+	pty    ptyAllocation
+	mu     sync.Mutex
+	closed bool
+}
+
+func (s *platformSession) Close() error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	if s.closed {
+		return nil
+	}
+	s.closed = true
+	if s.pty == nil {
+		return nil
+	}
+	return s.pty.Close()
+}
+
+func (s *platformSession) PTYKind() PTYKind {
+	if s == nil {
+		return PTYKindUnknown
+	}
+	return s.kind
+}
+
+func newPlatformSession(launch ProcessLaunch, cfg SessionConfig, kind PTYKind, pty ptyAllocation) (*platformSession, error) {
+	if pty == nil {
+		return nil, errors.New("agypty: PTY allocation is required")
+	}
+	return &platformSession{
+		launch: launch,
+		cfg:    cfg,
+		kind:   kind,
+		pty:    pty,
+	}, nil
+}
+
+type sessionProcess struct {
+	cmd       *exec.Cmd
+	tree      process.SubprocessTree
+	winHandle uintptr
+	exitCode  int
+}
+
+func (p *sessionProcess) Terminate() error {
+	if p == nil || p.cmd == nil {
+		return nil
+	}
+	return process.TerminateSubprocessTree(p.cmd, p.tree)
+}
+
+func (p *sessionProcess) Close() {
+	if p == nil || p.cmd == nil {
+		return
+	}
+	process.CloseSubprocessTree(p.cmd, p.tree)
+}
+
+func (p *sessionProcess) PID() int {
+	if p == nil || p.cmd == nil || p.cmd.Process == nil {
+		return 0
+	}
+	return p.cmd.Process.Pid
+}
+
+func exitCodeFromWait(waitErr error, proc *sessionProcess) int {
+	if waitErr != nil {
+		return -1
+	}
+	if proc != nil {
+		return proc.exitCode
+	}
+	return 0
+}
 
 func (s *platformSession) Run(ctx context.Context) (SessionResult, error) {
 	s.mu.Lock()
