--- pkg/workers/agypty/session_run_test_support.go
@@ -1,15 +1,0 @@
-package agypty
-
-import (
-	"os/exec"
-
-	"github.com/portpowered/infinite-you/pkg/workers/process"
-)
-
-func processAttachForTest(cmd *exec.Cmd) (process.SubprocessTree, error) {
-	return process.AttachSubprocessTree(cmd)
-}
-
-func processConfigureForTest(cmd *exec.Cmd) {
-	process.ConfigureSubprocessTree(cmd)
-}
